+++ src/main/java/com/heanoria/tests/banking/core/domain/operations/OperationAction.java
@@ -3,5 +3,5 @@ package com.heanoria.tests.banking.core.domain.operations;
 import com.heanoria.tests.banking.core.domain.data.OperationDetails;
 
 public interface OperationAction {
-    OperationDetails perform(double amount, double lastBalance);
+    OperationDetails perform(int amount, int lastBalance);
 }
